build-clang: Only support master + the latest stable version
My only motivation for this script is building kernels. Clang 6.0 was
the first version to support this without any hacks, this is the
earliest version that people should be building with this script.

Currently, Clang 6.0 is the stable branch and Clang 7.0 is master. In
the future, these versions will be bumped as development progresses.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: khusika <khusikadhamar@gmail.com>